fix: إصلاح خطأ نوع البيانات في حساب النصيب - إضافة toDecimal() لتحويل النتيجة

+++ lib/screens/reports/supplier_details_report_screen.dart
@@ -811,7 +811,7 @@ class _SupplierDetailsReportScreenState
                           Text(
                             formatCurrency(
                               partnerName != null && sharePercentage != null
-                                  ? (widget.totalProfit * sharePercentage / Decimal.fromInt(100))
+                                  ? (widget.totalProfit * sharePercentage / Decimal.fromInt(100)).toDecimal()
                                   : widget.totalProfit
                             ),
                             style: const TextStyle(
